Started refactoring to phpDocumentor/Guides
This is the start of our effort to adopt the work done by Doctrine but
in a way that it is easier to integrate into phpDocumentor; including
our own template system and to split it so that it can be more easily
cached.

I have decided to restructure the project towards our goals and with the
phpDocumentor namespaces; but I need to find a way to describe better
that the roots for this project is in Doctrine/rst-parser and ultimately
Gregwar/rst-parser.